[Edited]: move patches
