+++ lpp/patches.txt
@@ -5,4 +5,4 @@
 [post_model_sync]
 # Patches added in this section will be executed after doctypes are migrated
 
-lpp.patches.remove_field_from_item
+lpp.lpp.patches.remove_field_from_item
